change backend to use new structs

+++ backend/src/mandelbrot/Fractal.cpp
@@ -1,14 +1,14 @@
 #include "Fractal.h"
 
-const double boundX = 4;
-const double boundY = 4;
+const long double boundX = 4;
+const long double boundY = 4;
 
 double Fractal::deltaReal(double maxReal, double minReal, int xRes) {
-	return (maxReal - minReal) / xRes;
+    return (maxReal - minReal) / xRes;
 }
 
 double Fractal::deltaImaginary(double maxImaginary, double minImaginary, int yRes) {
-	return (maxImaginary - minImaginary) / yRes;
+    return (maxImaginary - minImaginary) / yRes;
 }
